add a delegating streamMessage for dedup
