+++ core/src/main/java/com/linecorp/armeria/client/DecodedHttpResponse.java
@@ -16,34 +16,27 @@
 
 package com.linecorp.armeria.client;
 
-import java.util.concurrent.CompletableFuture;
-
-import org.reactivestreams.Subscriber;
-
 import com.linecorp.armeria.common.HttpData;
 import com.linecorp.armeria.common.HttpObject;
 import com.linecorp.armeria.common.annotation.Nullable;
-import com.linecorp.armeria.common.stream.StreamCallbackListener;
+import com.linecorp.armeria.common.stream.DelegatingStreamMessage;
 import com.linecorp.armeria.common.stream.StreamMessageAndWriter;
-import com.linecorp.armeria.common.stream.SubscriptionOption;
 import com.linecorp.armeria.internal.common.InboundTrafficController;
 
 import io.netty.channel.EventLoop;
 import io.netty.util.concurrent.EventExecutor;
 
-final class DecodedHttpResponse implements DecodedHttpResponseWriter,
-                                           StreamCallbackListener<HttpObject> {
+final class DecodedHttpResponse extends DelegatingStreamMessage<HttpObject>
+        implements DecodedHttpResponseWriter {
 
     private final EventLoop eventLoop;
-    private final StreamMessageAndWriter<HttpObject> delegate;
     @Nullable
     private InboundTrafficController inboundTrafficController;
     private long writtenBytes;
 
     DecodedHttpResponse(EventLoop eventLoop, StreamMessageAndWriter<HttpObject> delegate) {
+        super(delegate);
         this.eventLoop = eventLoop;
-        this.delegate = delegate;
-        delegate.setCallbackListener(this);
     }
 
     @Override
@@ -62,23 +55,8 @@ final class DecodedHttpResponse implements DecodedHttpResponseWriter,
     }
 
     @Override
-    public void abort() {
-        delegate.abort();
-    }
-
-    @Override
-    public void abort(Throwable cause) {
-        delegate.abort(cause);
-    }
-
-    @Override
-    public boolean isOpen() {
-        return delegate.isOpen();
-    }
-
-    @Override
     public boolean tryWrite(HttpObject obj) {
-        final boolean published = delegate.tryWrite(obj);
+        final boolean published = delegate().tryWrite(obj);
         if (published && obj instanceof HttpData) {
             final int length = ((HttpData) obj).length();
             assert inboundTrafficController != null;
@@ -89,47 +67,11 @@ final class DecodedHttpResponse implements DecodedHttpResponseWriter,
     }
 
     @Override
-    public CompletableFuture<Void> whenConsumed() {
-        return delegate.whenConsumed();
-    }
-
-    @Override
-    public void close() {
-        delegate.close();
-    }
-
-    @Override
-    public void close(Throwable cause) {
-        delegate.close(cause);
-    }
-
-    @Override
     public void onRemoval(HttpObject obj) {
         if (obj instanceof HttpData) {
             final int length = ((HttpData) obj).length();
             assert inboundTrafficController != null;
             inboundTrafficController.dec(length);
         }
-    }
-
-    @Override
-    public boolean isEmpty() {
-        return delegate.isEmpty();
-    }
-
-    @Override
-    public long demand() {
-        return delegate.demand();
-    }
-
-    @Override
-    public CompletableFuture<Void> whenComplete() {
-        return delegate.whenComplete();
-    }
-
-    @Override
-    public void subscribe(Subscriber<? super HttpObject> subscriber, EventExecutor executor,
-                          SubscriptionOption... options) {
-        delegate.subscribe(subscriber, executor, options);
     }
 }
